requirements.txt
streamlit
openai
feedparser
python-dotenv
fpdf